+++ packages/sdk/src/evm/common/currency/isNativeToken.ts
@@ -1,0 +1,9 @@
+import { NATIVE_TOKEN_ADDRESS } from "../../constants/currency";
+import { constants } from "ethers";
+
+export function isNativeToken(tokenAddress: string): boolean {
+  return (
+    tokenAddress.toLowerCase() === NATIVE_TOKEN_ADDRESS ||
+    tokenAddress.toLowerCase() === constants.AddressZero
+  );
+}
