+++ packages/sdk/src/evm/common/currency/toDisplayValue.ts
@@ -1,0 +1,8 @@
+import { BigNumberish, utils } from "ethers";
+
+export function toDisplayValue(
+  amount: BigNumberish,
+  decimals: BigNumberish,
+): string {
+  return utils.formatUnits(amount, decimals);
+}
