+++ packages/sdk/src/evm/common/currency/toEther.ts
@@ -1,0 +1,5 @@
+import { BigNumberish, utils } from "ethers";
+
+export function toEther(amount: BigNumberish): string {
+  return utils.formatEther(amount);
+}
