+++ packages/sdk/src/evm/common/currency/toWei.ts
@@ -1,0 +1,7 @@
+import { AmountSchema } from "../../../core/schema/shared";
+import { Amount } from "../../types/currency";
+import { BigNumber, utils } from "ethers";
+
+export function toWei(amount: Amount): BigNumber {
+  return utils.parseEther(AmountSchema.parse(amount));
+}
