+++ packages/sdk/src/evm/common/marketplace.ts
@@ -11,7 +11,7 @@ import {
   Offer,
   UnmappedOffer,
 } from "../types/marketplace";
-import { fetchCurrencyValue } from "./currency";
+import { fetchCurrencyValue } from "./currency/fetchCurrencyValue";
 import type { IERC1155, IERC165, IERC721 } from "@thirdweb-dev/contracts-js";
 import ERC165Abi from "@thirdweb-dev/contracts-js/dist/abis/IERC165.json";
 import ERC721Abi from "@thirdweb-dev/contracts-js/dist/abis/IERC721.json";
