+++ packages/sdk/src/evm/contracts/prebuilt-implementations/vote.ts
@@ -1,7 +1,3 @@
-import {
-  fetchCurrencyMetadata,
-  fetchCurrencyValue,
-} from "../../common/currency";
 import { resolveAddress } from "../../common/ens";
 import { buildTransactionFunction } from "../../common/transactions";
 import { ContractAppURI } from "../../core/classes/contract-appuri";
@@ -37,6 +33,8 @@ import {
   Contract,
   ethers,
 } from "ethers";
+import { fetchCurrencyMetadata } from "../../common/currency/fetchCurrencyMetadata";
+import { fetchCurrencyValue } from "../../common/currency/fetchCurrencyValue";
 
 /**
  * Create a decentralized organization for token holders to vote on proposals.
